+++ udata/core/contact_point/api_fields.py
@@ -2,7 +2,7 @@ from udata.api import api, fields
 from udata.core.organization.api_fields import org_ref_fields
 from udata.core.user.api_fields import user_ref_fields
 
-DEFAULT_MASK = ",".join(("id", "name", "email", "contact_form"))
+DEFAULT_MASK = ",".join(("id", "name", "email", "contact_form", "role"))
 
 contact_point_fields = api.model(
     "ContactPoint",
@@ -17,6 +17,7 @@ contact_point_fields = api.model(
         "owner": fields.Nested(
             user_ref_fields, allow_null=True, description="The user information"
         ),
+        "role": fields.String(description="The role of the contact", required=True),
     },
     mask=DEFAULT_MASK,
 )