+++ udata/tests/test_rdf.py
@@ -7,10 +7,12 @@ from rdflib import (
 from udata.models import ContactPoint
 from udata.rdf import (
     ACCEPTED_MIME_TYPES,
+    DCAT,
+    DCT,
     FORMAT_MAP,
     RDF,
     VCARD,
-    contact_point_to_rdf,
+    contact_points_to_rdf,
     guess_format,
     negociate_content,
     want_rdf,
@@ -82,16 +84,48 @@ class GuessFormatTest(object):
 
 
 class ContactToRdfTest:
-    def test_contact_point_to_rdf(self):
+    def test_contact_points_to_rdf(self):
         contact = ContactPoint(
             name="Organization contact",
             email="hello@its.me",
             contact_form="https://data.support.com",
         )
 
-        contact_rdf = contact_point_to_rdf(contact, None)
+        contact_rdfs = contact_points_to_rdf([contact], None)
 
-        assert contact_rdf.value(RDF.type).identifier == VCARD.Kind
-        assert contact_rdf.value(VCARD.fn) == Literal("Organization contact")
-        assert contact_rdf.value(VCARD.hasEmail).identifier == URIRef("mailto:hello@its.me")
-        assert contact_rdf.value(VCARD.hasUrl).identifier == URIRef("https://data.support.com")
+        for contact_point, predicate in contact_rdfs:
+            assert contact_point.value(RDF.type).identifier == VCARD.Kind
+            assert contact_point.value(VCARD.fn) == Literal("Organization contact")
+            assert contact_point.value(VCARD.hasEmail).identifier == URIRef("mailto:hello@its.me")
+            assert contact_point.value(VCARD.hasUrl).identifier == URIRef(
+                "https://data.support.com"
+            )
+            # Default predicate is "contact"
+            assert predicate == DCAT.contactPoint
+
+    @pytest.mark.parametrize(
+        "role,predicate",
+        [
+            ("contact", DCAT.contactPoint),
+            ("publisher", DCT.publisher),
+            ("creator", DCT.creator),
+        ],
+    )
+    def test_contact_points_to_rdf_roles(self, role, predicate):
+        contact = ContactPoint(
+            name="Organization contact",
+            email="hello@its.me",
+            contact_form="https://data.support.com",
+            role=role,
+        )
+
+        contact_rdfs = contact_points_to_rdf([contact], None)
+
+        for contact_point, contact_point_predicate in contact_rdfs:
+            assert contact_point.value(RDF.type).identifier == VCARD.Kind
+            assert contact_point.value(VCARD.fn) == Literal("Organization contact")
+            assert contact_point.value(VCARD.hasEmail).identifier == URIRef("mailto:hello@its.me")
+            assert contact_point.value(VCARD.hasUrl).identifier == URIRef(
+                "https://data.support.com"
+            )
+            assert contact_point_predicate == predicate