+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Add search to posts [#3248](https://github.com/opendatateam/udata/pull/3248)
 
 ## 10.0.7 (2025-01-13)