+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current (in progress)
 
 - Harvest attribution roles [#3149](https://github.com/opendatateam/udata/pull/3149)
+- Resource catalog: fix preview_url and add extras [#3188](https://github.com/opendatateam/udata/pull/3188)
+- Add trailing slash to security routes [#3251](https://github.com/opendatateam/udata/pull/3251)
 
 ## 10.0.7 (2025-01-13)