+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Resource catalog: fix preview_url and add extras [#3188](https://github.com/opendatateam/udata/pull/3188)
 - Add trailing slash to security routes [#3251](https://github.com/opendatateam/udata/pull/3251)
 - Upgrade packaging dependency to 24.2 [#3252](https://github.com/opendatateam/udata/pull/3252)
+- Fix contact point permissinos [#3253](https://github.com/opendatateam/udata/pull/3253)
 
 ## 10.0.7 (2025-01-13)